Extract news logic into hook
