+++ src/screens/news/NewsScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import {
   FlatList,
   Image,
@@ -8,79 +8,28 @@ import {
   View,
 } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
-import NewsRepository from '../../data/NewsRepository'
-import ProfileRepository from '../../data/ProfileRepository'
 import { Colors, Spacing } from '../../styles'
 import { Analytics } from '../../utils/Analytics'
 import { ExternalLink } from '../shared/ExternalLink'
 import LoaderView from '../shared/LoaderView'
-import { StatefulView, ViewState } from '../shared/StatefulView'
-import { ViewStateUtils } from '../shared/ViewStateUtils'
+import { StatefulView } from '../shared/StatefulView'
 import NewsContentViewModel from './NewsContentViewModel'
-import { NewsContentViewModelMapper } from './NewsContentViewModelMapper'
 import NewsRow from './NewsRow'
 import { NewsRowViewModel } from './NewsRowViewModel'
+import { useNewsScreen } from './useNewsScreen.hook'
 
 const Separator = () => {
   return <View style={styles.separator} />
 }
 
 const NewsScreen = () => {
-  const [statefulState, setStatefulState] = useState<
-    ViewState.Type<NewsContentViewModel>
-  >(new ViewState.Loading())
-  const [isRefreshing, setRefreshing] = useState(true)
-  const [isLoadingMore, setLoadingMore] = useState(false)
-  const fetchNews = async (page: number) => {
-    const zipCode = await ProfileRepository.getInstance().getZipCode()
-    return NewsRepository.getInstance().getNews(zipCode, page)
-  }
-
-  const loadFirstPage = () => {
-    setRefreshing(true)
-    fetchNews(1)
-      .then((paginatedResult) => {
-        const content = NewsContentViewModelMapper.map(paginatedResult)
-        setStatefulState(new ViewState.Content(content))
-      })
-      .catch((error) => {
-        setStatefulState(
-          ViewStateUtils.networkError(error, () => {
-            setStatefulState(new ViewState.Loading())
-            loadFirstPage()
-          }),
-        )
-      })
-      .finally(() => setRefreshing(false))
-  }
-
-  const loadMore = () => {
-    const currentState = statefulState
-    if (currentState instanceof ViewState.Content) {
-      const content = currentState.content
-      const paginationInfo = content.paginationInfo
-      if (paginationInfo.currentPage === paginationInfo.lastPage) {
-        // last page reached : nothing to paginate
-        return
-      }
-      setLoadingMore(true)
-      fetchNews(paginationInfo.currentPage + 1)
-        .then((paginatedResult) => {
-          const newContent = NewsContentViewModelMapper.map(
-            paginatedResult,
-            content,
-          )
-          setStatefulState(new ViewState.Content(newContent))
-        })
-        .catch((error) => {
-          console.log(error)
-          // no-op: next page can be reloaded by reaching the end of the list again
-        })
-        .finally(() => setLoadingMore(false))
-    }
-  }
-
-  useEffect(loadFirstPage, [])
+  const {
+    statefulState,
+    isLoadingMore,
+    isRefreshing,
+    loadFirstPage,
+    loadMore,
+  } = useNewsScreen()
 
   const renderItem = ({ item }: ListRenderItemInfo<NewsRowViewModel>) => {
     return (
